feat: implement header for checkout page
